i2c_oled/grok/requirements.txt
luma.core <= 2.4.2
luma.oled <= 3.14.0
